refactor(ripgrep): extract streamLines to io.ts and rename ensureInstalled to rg
